fonction hsow

+++ src/Controller/TrajetsController.php
@@ -23,6 +23,7 @@ use Symfony\Component\Form\FormError;
 class TrajetsController extends AbstractController
 {
     
+
     #[Route('/', name: 'app_trajets_index', methods: ['GET'])]
     #[IsGranted('ROLE_USER')]
     public function index(TrajetsRepository $trajetsRepository): Response
@@ -49,11 +50,12 @@ class TrajetsController extends AbstractController
 
         
         if ($form->isSubmitted() && $form->isValid()) {
+           
             //On récupère les données du formulaire
             $trajet = $form->getData();
 
             //On vérifie d'abord si les villes existent déjà dans la base de donnée
-            // var_dump($form->get('demarrea')->getData());
+
             $villeDepart = $manager->getRepository(Villes::class)->find(['id' => $form->getData()->getDemarreA()]);
             $villeArrivee = $manager->getRepository(Villes::class)->find(['id' => $form->getData()->getArriveA()]);
             $trajet->setArriveA($villeArrivee);
@@ -81,6 +83,8 @@ class TrajetsController extends AbstractController
                 }
             }
 
+
+           
             // champs remplis d'office:
             $trajet->setPublie($this->getUser());
             $trajet->setEtat('ouvert');
@@ -104,15 +108,66 @@ class TrajetsController extends AbstractController
         ]);
     }
     
+
+    #[Route('/{id}', name: 'app_trajets_show', methods: ['GET'])]
+    public function show(Trajets $trajet, Request $request, EntityManagerInterface $manager): Response
+    {
+        // modifications automatiques de l'état d'un trajet
+        // dans l'affichage
+        $demain = new DateTime('tomorrow');
+        if ($trajet->getTDepart() <$demain ) {
+            $trajet->setEtat('bloqué');
+
+            $this->addFlash(
+                'succès',
+                'Votre trajet ne peut plus être modifié !'
+            );    
+        }
+        $maintenant = new DateTime();
+        $hier = new DateTime('yesterday');
+        if ($trajet->getTArrivee() <$maintenant  || $trajet->getTDepart() <$hier ) {
+            $trajet->setEtat('terminé');
+            $this->addFlash(
+                'succès',
+                'Votre trajet est terminé !'
+            );    
+        }
+        
+        $manager->persist($trajet);
+
+        $manager->flush();
+
+        return $this->render('trajets/show.html.twig', [
+            'trajet' => $trajet,
+        ]);
+    }
+
+    
     #[Route('/{id}/edit', name: 'app_trajets_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Trajets $trajet, TrajetsRepository $trajetsRepository): Response
+    public function edit(Request $request, Trajets $trajet, TrajetsRepository $trajetsRepository, EntityManagerInterface $manager): Response
     {
         $form = $this->createForm(TrajetsType::class, $trajet);
         $form->handleRequest($request);
+        $trajet = $form->getData();
+        
+        $demain = new DateTime('tomorrow');
+        if ($trajet->getTDepart() <$demain ) {
+            $this->addFlash(
+                'warning',
+                'Vous ne pouvez plus modifier ce trajet.'
+            );
+            
+            return $this->redirectToRoute('app_trajets_index');
+        }
+
+
 
         if ($form->isSubmitted() && $form->isValid()) {
             $trajetsRepository->save($trajet, true);
+            
+            $manager->persist($trajet);
 
+            $manager->flush();
             return $this->redirectToRoute('app_trajets_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -125,12 +180,27 @@ class TrajetsController extends AbstractController
 
     
     #[Route('/{id}', name: 'app_trajets_delete', methods: ['POST'])]
-    public function delete(Request $request, Trajets $trajet, TrajetsRepository $trajetsRepository): Response
+    public function delete(Request $request, Trajets $trajet, TrajetsRepository $trajetsRepository, EntityManagerInterface $manager): Response
     {
+        $demain = new DateTime('tomorrow');
+        if ($trajet->getTDepart() <$demain ) {
+            $trajet->setEtat('bloqué');
+            $this->addFlash(
+                'warning',
+                'Vous ne pouvez plus supprimer ce trajet.'
+            );
+
+            return $this->redirectToRoute('app_trajets_index');
+        }
+        
         if ($this->isCsrfTokenValid('delete'.$trajet->getId(), $request->request->get('_token'))) {
+            $trajet->setEtat('annulé');
+            // si on l'enlève carrément:
             $trajetsRepository->remove($trajet, true);
         }
+        $manager->persist($trajet);
 
+        $manager->flush();
         return $this->redirectToRoute('app_trajets_index', [], Response::HTTP_SEE_OTHER);
     }
 
@@ -173,6 +243,5 @@ class TrajetsController extends AbstractController
             ]);
         }
     }
-    
     
 }
